+++ src/lib/application.ts
@@ -33,6 +33,7 @@ import {
     getWatchEntryPoints,
 } from "./utils/entry-point";
 import { nicePath } from "./utils/paths";
+import { hasBeenLoadedMultipleTimes } from "./utils/general";
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const packageInfo = require("../../package.json") as {
@@ -149,6 +150,12 @@ export class Application extends ChildableComponent<
             }
         }
         this.options.read(this.logger);
+
+        if (hasBeenLoadedMultipleTimes()) {
+            this.logger.warn(
+                `TypeDoc has been loaded multiple times. This is commonly caused by plugins which have their own installation of TypeDoc. This will likely break things.`
+            );
+        }
     }
 
     /**